ignoreIP

+++ back/src/common/remove-remote-user.js
@@ -1,18 +1,20 @@
 import request from "request";
 
-export default function removeRemoteUser(id) {
+export default function removeRemoteUser(id, ignoreIP) {
   global.OTHER_IPS.forEach((ip) => {
-    request.delete(
-      {
-        url: `http://${ip}:${process.env.SERVER_PORT}/sync/delete/${id}`,
-      },
-      (err) => {
-        if (err) {
-          console.log(err);
-        } else {
-          console.log("sync remove success");
+    if (ip !== ignoreIP) {
+      request.delete(
+        {
+          url: `http://${ip}:${process.env.SERVER_PORT}/sync/delete/${id}`,
+        },
+        (err) => {
+          if (err) {
+            console.log(err);
+          } else {
+            console.log("sync remove success");
+          }
         }
-      }
-    );
+      );
+    }
   });
 }
